% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benford_table_functions.R
\name{Benford_table}
\alias{Benford_table}
\title{Generates contingency table for Benford distribution (a.k.a expected digit frequency under Benford's Law)}
\usage{
Benford_table(N, out_fp = NA, save = FALSE)
}
\arguments{
\item{N}{Number of digits to generate}

\item{out_fp}{Filepath to save the table to. Default to NA.}

\item{save}{TRUE or FALSE: saving the table or not. Default to FALSE.}
}
\value{
Benford contingency table generated.
}
\description{
Generates contingency table for Benford distribution (a.k.a expected digit frequency under Benford's Law)
}
