% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_low_test.R
\name{calculate_weighted_p}
\alias{calculate_weighted_p}
\title{Computes weighted average probability of high digits across all digit places.
Helper function for \code{high_low_by_digit_place}.}
\usage{
calculate_weighted_p(high_and_low_total_counts, high_freq_theoratical)
}
\arguments{
\item{high_freq_theoratical}{A table for theoratical high digit frequency in each digit place.}
}
\value{
Weighted average probability of high digits across all digit places.
}
\description{
Computes weighted average probability of high digits across all digit places.
Helper function for \code{high_low_by_digit_place}.
}
