% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_goodness_of_fit_functions.R
\name{chi_square_gof}
\alias{chi_square_gof}
\title{Perform chi square test for goodness of fit test}
\usage{
chi_square_gof(observed_table, expected_table, df, freq = TRUE)
}
\arguments{
\item{observed_table}{Observation table for chi square test}

\item{expected_table}{Expected table for chi square test}

\item{df}{Degrees of freedom}

\item{freq}{TRUE or FALSE: Whether the input expected table is in decimal.}
}
\value{
p-value for chi square goodness of fit test
}
\description{
Perform chi square test for goodness of fit test
}
