% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_test.R
\name{compute_percent_rounded_digits}
\alias{compute_percent_rounded_digits}
\title{Compute the percent of rounded digits in a given dataframe.
Removes all 0 entries before begin.}
\usage{
compute_percent_rounded_digits(data, rounding_patterns)
}
\arguments{
\item{data}{The dataframe to analyze percent rounded digits on}

\item{rounding_patterns}{The patterns to be counted as rounding digits.
\itemize{
  \item An array of characters such as c('0','00','000','5','50','500', '75', '25').
  \item Defaulted to c('0','00','000','0000', '00000', '000000', '5', '50', '500').
  \item \code{n_zeros_pattern} might be helpful for generating strings of 0s.
}}
}
\value{
The percentage of rounded digits in input data
}
\description{
Compute the percent of rounded digits in a given dataframe.
Removes all 0 entries before begin.
}
