% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_test.R
\name{find_percent_repeats}
\alias{find_percent_repeats}
\title{Finds the percent of repeats in the given data based on given definition of a repeat (what columns need to match)}
\usage{
find_percent_repeats(data, duplicate_matching_cols)
}
\arguments{
\item{duplicate_matching_cols}{The names of data columns two rows need to match exactly in order to be defined as a repeat}

\item{dataThe}{The dataframe to compute percent repeat on}
}
\value{
The percentage of repeated numbers in input data
}
\description{
Finds the percent of repeats in the given data based on given definition of a repeat (what columns need to match)
}
