% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_goodness_of_fit_functions.R
\name{get_df}
\alias{get_df}
\title{Compute the degrees of freedom}
\usage{
get_df(table, standard = FALSE)
}
\arguments{
\item{standard}{}
}
\value{

}
\description{
Compute the degrees of freedom
}
