% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_round_numbers_test.R
\name{get_round_unround_digitdata}
\alias{get_round_unround_digitdata}
\title{Helper function for unpack_round_numbers_test.
Creates two two copies of digitdata:
one with only round entries and the other with only unround entries in the specified numeric column}
\usage{
get_round_unround_digitdata(digitdata, unpacking_rounding_column)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{unpacking_rounding_column}{The data column (numeric!) to split rounded and unrounded digits upon to perform unpacking rounding test.}
}
\value{
A list with two DigitAnalysis objects: round_digitdata and unround_digitdata
}
\description{
Helper function for unpack_round_numbers_test.
Creates two two copies of digitdata:
one with only round entries and the other with only unround entries in the specified numeric column
}
