% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{hist_2D}
\alias{hist_2D}
\title{Plot 2d histogram given data. Either rownames or colnames will be x values, and the data will be y values.}
\usage{
hist_2D(
  data,
  data_style = "row",
  xlab = "digits",
  ylab = "frequency",
  title = "2D Histogram",
  hline = NA
)
}
\arguments{
\item{data}{The 1D dataframe to be plotted.}

\item{data_style}{The style of input \code{data}. \code{ggplot} requires columns to be the arrays for x values, y values, etc.
\itemize{
  \item If \code{data} has a row for y values, pass in 'row'.
  \item If \code{data} has a column for y values, pass in 'col'.
  \item Defaulted to 'row'.
}}

\item{xlab}{x-axis label. Defaulted to 'digits'.}

\item{ylab}{y-axis label. Defaulted to 'frequency'.}

\item{title}{Plot title. Defaulted to '2D Histogram'.}

\item{hline}{Specifies the y-intercept value if a horizontal line is desired. Defaulted to NA.}
}
\value{
A ggplot instance.
}
\description{
Plot 2d histogram given data. Either rownames or colnames will be x values, and the data will be y values.
}
