% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_functions.R
\name{make_aligned_data}
\alias{make_aligned_data}
\title{Create the dataframe with the left/right aligned digits of each data column to be analyzed->'left_aligned'/'right_aligned' of the class}
\usage{
make_aligned_data(
  cleaned_data,
  col_analyzing,
  naming_method,
  align_direction = "left"
)
}
\arguments{
\item{cleaned_data}{'cleaned' slot for DigitAnalysis}

\item{col_analyzing}{The numeric columns to be used}

\item{naming_method}{Either \code{left_aligned_column_names} or \code{right_aligned_column_names} from DigitAnalysis class}

\item{align_direction}{'left' or 'right': create left-aligned or right-aligned dataframe}
}
\value{
The dataframe with the left/right aligned digits of each data column to be analyzed for DigitAnalysis slot 'left_aligned'/'right_aligned'
}
\description{
Create the dataframe with the left/right aligned digits of each data column to be analyzed->'left_aligned'/'right_aligned' of the class
}
