% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_functions.R
\name{make_class}
\alias{make_class}
\title{Create an object instance for \code{DigitAnalysis}. Parse and clean the data for digit analysis.}
\usage{
make_class(
  col_analyzing,
  filepath = NA,
  filetype = "csv",
  delim = ",",
  raw_df = NA
)
}
\arguments{
\item{col_analyzing}{All potential datra columns to be analyzed. Can be specified as any of
\itemize{
  \item \code{col_analyzing} = 'col_name'
  \item \code{col_analyzing} = c('col_name')
  \item \code{col_analyzing} = c('col_name1','col_name2', ...)
}}

\item{filepath}{Default to NA. If loading data using filepath, specify filepath as a string.}

\item{filetype}{Default to 'csv'. If loading data using filepath, specify either 'csv' or 'excel'.}

\item{delim}{Defaulted to ','. Can specify other delimeters as well.}

\item{raw_df}{Default to NA. If loading data using a dataframe. Pass in the dataframe instance.}
}
\value{
An object in \code{DigitAnalysis}
}
\description{
Create an object instance for \code{DigitAnalysis}. Parse and clean the data for digit analysis.
}
\examples{
make_class('col_name', filepath='~/filename.csv')
make_class('col_name', filepath='~/filename.xlsx', filetype='excel', delim=',')
make_class('col_name', raw_df=my_dataframe)
}
