% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_digit_test_helper_functions.R
\name{make_sub_digitdata}
\alias{make_sub_digitdata}
\title{#' Updates \code{grouping} to default if it is NA.
#'
#' @param grouping The grouping in \code{column}.
#' @param column The column to break data on
#' @inheritParams all_digits_test
#'
#' @return \code{grouping} that is not NA.
get_grouping = function(grouping, column, digitdata){
  if (is.na(grouping)){
    unique_items = unique(digitdata@cleaned[column])[[1]]
    grouping = list()
    for (item in unique_items){
      grouping[[item]] = item
    }
  }
  return(grouping)
}
Create a sub-object of the \code{digitdata} object given the indexes of the rows that belong to the sub-object.}
\usage{
make_sub_digitdata(digitdata, indexes)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{indexes}{The indexes of the rows that belong to the sub-object.}
}
\value{
A sub-object of input \code{digitdata}.
}
\description{
#' Updates \code{grouping} to default if it is NA.
#'
#' @param grouping The grouping in \code{column}.
#' @param column The column to break data on
#' @inheritParams all_digits_test
#'
#' @return \code{grouping} that is not NA.
get_grouping = function(grouping, column, digitdata){
  if (is.na(grouping)){
    unique_items = unique(digitdata@cleaned[column])[[1]]
    grouping = list()
    for (item in unique_items){
      grouping[[item]] = item
    }
  }
  return(grouping)
}
Create a sub-object of the \code{digitdata} object given the indexes of the rows that belong to the sub-object.
}
