% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_test.R
\name{n_zeros_pattern}
\alias{n_zeros_pattern}
\title{Gives an array of strings of zeros from length 1 to n.
Helper function for users inputing \code{rounding_patterns} for \code{rounding_test}.}
\usage{
n_zeros_pattern(n)
}
\arguments{
\item{n}{The maximum length of 0s string to generate.}
}
\value{
An array of strings of zeros from length 1 to n.
}
\description{
Gives an array of strings of zeros from length 1 to n.
Helper function for users inputing \code{rounding_patterns} for \code{rounding_test}.
}
\examples{
n_zeros_pattern(10)
}
