% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padding_test.R
\name{padding_test}
\alias{padding_test}
\title{Performs padding test vs stimulations of Benford conforming datasets via percentile}
\usage{
padding_test(
  digitdata,
  contingency_table,
  data_columns = "all",
  max_length = 8,
  num_digits = 5,
  N = 10000,
  omit_05 = c(0, 5),
  break_out = NA
)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{contingency_table}{The Benford probability table.}

\item{data_columns}{The names of numeric columns of data to be analyzed. Default can be 'all', where using all data columns in \code{numbers} df in \code{digitdata};
an array of column names, as characters; a single column name, as character.}

\item{max_length}{The length of the longest numbers considered. Defaulted to 8.}

\item{num_digits}{The total number of digits aligned from the right to be analyzed. Defaulted to 5, meaning analyzing digit place 1s to 10ks.}

\item{N}{The number of Benford conforming datasets to stimulate.}

\item{omit_05}{Whether to omit 0 or both 0 and 5. If omit both 0 and 5, pass in c(0,5) or c(5,0); if omit only 0 pass in 0 or c(0); if omit neither, pass in NA. Default to c(0,5).}

\item{break_out}{The data column (non-numeric!) to split up the dataset based on different categories in the column if specified as an character. Default to NA.}
}
\value{
\itemize{
  \item A list with 4 elements
  \itemize{
    \item \code{expected_mean}: the expected mean by Benford's Law
    \item \code{observed_mean}: the mean of the input data
    \item \code{diff_in_mean}: the mean difference betweeen observed_mean and expected_mean
    \item \code{p_values}: the percentile of the observed dataset among all stimulated datasets in decreasing order
  }
  \item Plots on \code{diff_in_mean} for each category if \code{plot = TRUE}
}
}
\description{
Performs padding test vs stimulations of Benford conforming datasets via percentile
}
\examples{
padding_test(digitdata, contingency_table, data_columns='all')
padding_test(digitdata, contingency_table, data_columns=c('col_name1', 'col_name2'), omit_05=NA)
padding_test(digitdata, contingency_table, data_columns='all', max_length=7, num_digits=3, omit_05=0)
padding_test(digitdata, contingency_table, data_columns='all', N=100, omit_05=NA, break_out='col_name')
}
