% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_test.R
\name{percent_repeats_by_sector}
\alias{percent_repeats_by_sector}
\title{Computes the percentage of repeated numbers for each sector}
\usage{
percent_repeats_by_sector(data, sector_column, sector_grouping)
}
\arguments{
\item{data}{The dataframe to compute percent repeat on}

\item{sector_column}{The column for splitting the data into sectors for separate analysis. Default to NA.
\itemize{
  \item If it is specified (not NA), then a sectorized test will be performed instead of a normal repeat test
}}

\item{sector_grouping}{A list of arrays
\itemize{
  \item Each the names of the elements in the list is the sector name
  \item Each array contains the categories belonging to that sector
}}
}
\value{
A table of percentage of repeated numbers for each sector
}
\description{
Computes the percentage of repeated numbers for each sector
}
