% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_multiple_hist2d}
\alias{plot_multiple_hist2d}
\title{Plot multiple plots on a single image}
\usage{
plot_multiple_hist2d(plot_list)
}
\arguments{
\item{plot_list}{A list of ggplot instances}
}
\value{
A plot instance with all plots in one single figure
}
\description{
Plot multiple plots on a single image
}
