% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_test.R
\name{sector_test}
\alias{sector_test}
\title{Performs sector test to analyze uneven distribution of percent repeats across sectors (supposed to be uniform).
A wrapper function for \code{repeat_test}.}
\usage{
sector_test(
  digitdata,
  sector_column,
  sector_grouping,
  duplicate_matching_cols = "all",
  break_out = NA,
  failure_factor = 3
)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{sector_column}{The column for splitting the data into sectors for separate analysis.}

\item{sector_grouping}{A list of arrays
\itemize{
  \item Each the names of the elements in the list is the sector name
  \item Each array contains the categories belonging to that sector
}}

\item{duplicate_matching_cols}{An array of names of data columns two rows need to match exactly in order to be defined as a repeat.
Default to 'all', meaning matching all columns in 'clean' slot of \code{digitdata}}

\item{break_out}{The data column (non-numeric!) to split up the dataset based on different categories in the column if specified as an character. Default to NA.}

\item{failure_factor}{NEED TO EDIT LATER}
}
\value{
\itemize{
  \item A table of p-values for repeat test by sector on each category
  \item Plots for each category with variables as sectors if \code{plot = TRUE}
  \item If NaN is in returned table, it means that there are no occurances of the data of the sector in that category --> 0/0 in percentage
}
}
\description{
Performs sector test to analyze uneven distribution of percent repeats across sectors (supposed to be uniform).
A wrapper function for \code{repeat_test}.
}
\examples{
sector_test(digitdata, sector_column='sector_name', sector_grouping=list('sector 1'=c('a'), 'sector 2'=c('b', 'c')))
sector_test(digitdata, sector_column='sector_name', sector_grouping=list('sector 1'=c('a, b'), 'sector 2'=c('c', 'd')),
duplicate_matching_cols=c('col_name1, col_name2'), break_out='col_name', failure_factor=3)
}
